+++ arviz/plots/posteriorplot.py
@@ -17,6 +17,7 @@ from ..rcparams import rcParams
 def plot_posterior(
     data,
     var_names=None,
+    transform=None,
     coords=None,
     figsize=None,
     textsize=None,
@@ -45,6 +46,8 @@ def plot_posterior(
         Refer to documentation of az.convert_to_dataset for details
     var_names : list of variable names
         Variables to be plotted, two variables are required.
+    transform : callable
+        Function to transform data (defaults to None i.e.the identity function)
     coords : mapping, optional
         Coordinates of var_names to be plotted. Passed to `Dataset.sel`
     figsize : tuple
@@ -177,6 +180,8 @@ def plot_posterior(
         >>> az.plot_posterior(data, var_names=['mu'], credible_interval=.75)
     """
     data = convert_to_dataset(data, group=group)
+    if transform is not None:
+        data = transform(data)
     var_names = _var_names(var_names, data)
 
     if coords is None: